+++ packages/reactive_ble_mobile/test/converter/protobuf_converter_test.dart
@@ -12,6 +12,7 @@ void main() {
     group("decoding ${pb.DeviceScanInfo}", () {
       const id = 'id';
       const name = 'name';
+      const connectable = Connectable.available;
 
       late pb.ServiceDataEntry serviceDataEntry1;
       late pb.ServiceDataEntry serviceDataEntry2;
@@ -35,6 +36,7 @@ void main() {
         message = pb.DeviceScanInfo()
           ..id = id
           ..name = name
+          ..isConnectable = (pb.IsConnectable()..code = 2)
           ..serviceData.add(serviceDataEntry1)
           ..serviceData.add(serviceDataEntry2)
           ..serviceUuids.add(serviceUuid1)
@@ -56,6 +58,16 @@ void main() {
             scanresult.result
                 .iif(success: (d) => d.name, failure: (_) => throw Exception()),
             name);
+      });
+
+      test('converts connectable', () {
+        expect(
+          scanresult.result.iif(
+            success: (d) => d.connectable,
+            failure: (_) => throw Exception(),
+          ),
+          connectable,
+        );
       });
 
       test('converts service data', () {
@@ -492,19 +504,23 @@ void main() {
 
           final discoveredInternalServices = pb.DiscoveredService()
             ..serviceUuid = internalServiceUuid
-            ..characteristics.add(pb.DiscoveredCharacteristic(
-                characteristicId: internalCharUuid,
-                serviceId: internalServiceUuid,
-                isReadable: true))
+            ..characteristics.add(
+              pb.DiscoveredCharacteristic()
+                ..characteristicId = internalCharUuid
+                ..serviceId = internalServiceUuid
+                ..isReadable = true,
+            )
             ..characteristicUuids.add(internalCharUuid);
 
           final discoveredService = pb.DiscoveredService()
             ..serviceUuid = serviceUuid
             ..characteristicUuids.add(charUuid)
-            ..characteristics.add(pb.DiscoveredCharacteristic(
-                characteristicId: charUuid,
-                serviceId: serviceUuid,
-                isWritableWithResponse: true))
+            ..characteristics.add(
+              pb.DiscoveredCharacteristic()
+                ..characteristicId = charUuid
+                ..serviceId = serviceUuid
+                ..isWritableWithResponse = true,
+            )
             ..includedServices.add(discoveredInternalServices);
 
           message = pb.DiscoverServicesInfo()