+++ packages/reactive_ble_mobile/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 5.1.1
+
+* Make Connectable backwards compatible #757, #750
+
+## 5.1.0
+
+* Add IsConnectable to discovery data. #750
+* Upgraded build_runner. #750
+
 ## 5.0.3
 
 * Enable extended advertising on android. Fix #571